Linked code to UI trigger. works now without logging.
