+++ appMetaFetcher/lib/getAppMetadata.js
