+++ appMetaFetcher/lib/getCustomPropertiesForEntity.js
