+++ appMetaFetcher/lib/getCustomPropertyDefinitions.js
