+++ appMetaFetcher/lib/getMeasureData.js
