+++ appMetaFetcher/lib/getNonMasterItemMetrics.js
