+++ appMetaFetcher/lib/getSheetData.js
