+++ appMetaFetcher/lib/getVariableData.js
