+++ appMetaFetcher/lib/getVisualizationData.js
