+++ appMetaFetcher/lib/stringExtensions.js
