+++ appMetaFetcher/lib/writeCSV.js
