+++ appMetaFetcher/lib/writeHeaders.js
